+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.4.6] - 2025-08-25
+### Fixed
+- Text and translation corrections in the fast payments flow
+- Responsiveness improvements in the fast payments flow
+- Fixed payment ID storage errors in custom fields metadata for Ticket and PSE payments
+- Fixed sync button functionality for refund scenarios failing to update order information via notification API
+### Changed
+- Changed installments selector for bank interests hint
+- Changed installments options for correctly translated texts on installments without fee
+
 ## [8.4.5] - 2025-08-18
 ### Fixed
 - Changed the error message when closing the modal in the fast payments flow